+++ Backend/Paybble/Core/Paybbles.Application/Features/Expenses/Commands/CreateExpense/CreateExpenseDTO.cs
@@ -5,7 +5,7 @@ namespace Paybble.Application.Features.Expenses .Commands.CreateExpense
     {
         public int Id { get; private set; }
         public string Description { get; private set; } = string.Empty;
-        public int Value { get; private set; }
+        public decimal Value { get; private set; }
         public bool Paid { get; private set; }
         public DateOnly DueDate { get; private set; }
     }
