+++ Backend/Paybble/Core/Paybbles.Application/Features/Expenses/Commands/UpdateExpense/UpdateExpenseCommand.cs
@@ -3,5 +3,12 @@ using Paybble.Domain.Enums;
 
 namespace Paybble.Application.Features.Expenses.Commands.UpdateExpense
 {
-    public record UpdateExpenseCommand(int id, string description, int value, DateOnly DueDate, Recurrence recurrence, int frequency) : IRequest;
+    public record UpdateExpenseCommand(
+        int id, 
+        string description, 
+        decimal value, 
+        DateOnly DueDate, 
+        Recurrence recurrence, 
+        int frequency
+    ) : IRequest;
 }
