+++ Backend/Paybble/Core/Paybbles.Application/Features/Incomes/Commands/CreateIncome/CreateIncomeDTO.cs
@@ -6,7 +6,7 @@ namespace Paybble.Application.Features.Incomes.Commands.CreateIncome
     {
         public int Id { get; private set; }
         public string Description { get; private set; } = string.Empty;
-        public int Value { get; private set; }
+        public decimal Value { get; private set; }
         public DateOnly? PaymentDate { get; set; }
     }
 }
