+++ Backend/Paybble/Core/Paybbles.Application/Features/Incomes/Commands/UpdateIncome/UpdateIncomeCommand.cs
@@ -3,5 +3,12 @@ using Paybble.Domain.Enums;
 
 namespace Paybble.Application.Features.Incomes.Commands.UpdateIncome
 {
-    public record UpdateIncomeCommand(int id, string description, int value, Recurrence recurrence, int frequency, DateOnly? PaymentDate) : IRequest;
+    public record UpdateIncomeCommand(
+        int id, 
+        string description,
+        decimal value, 
+        Recurrence recurrence, 
+        int frequency, 
+        DateOnly? PaymentDate
+    ) : IRequest;
 }
