+++ Backend/Paybble/Core/Paybbles.Application/Features/Incomes/Queries/GetIncomesByYearMonth/IncomesByYearMonthVm.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; private set; }
         public string Description { get; private set; } = string.Empty;
-        public int Value { get; private set; }
+        public decimal Value { get; private set; }
         public DateOnly PaymentDate { get; set; }
     }
 }
