+++ Backend/Paybble/Core/Paybbles.Application/Features/Savings/Commands/CreateSavings/CreateSavingsDTO.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Description { get; set; } = string.Empty;
-        public int GoalValue { get; set; }
+        public decimal GoalValue { get; set; }
         public int Progress { get; set; }
     }
 }
